[proto] Fixed fill type in AA
